requirements.txt
# Django and core dependencies
asgiref==3.9.1
Django==5.2.4
sqlparse==0.5.3

# REST Framework and JWT
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0

# CORS and middleware
django-cors-headers==4.7.0

# Image processing
pillow==11.3.0
pdf2image==1.17.0

# Jalali date support
django-jalali-date==2.5.0

# Environment variables
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0

# Additional utilities
pathlib2==2.3.7
